Fixing according to security refactoring
