+++ src/main/java/eu/h2020/symbiote/model/AcquisitionTask.java
@@ -5,15 +5,9 @@ import eu.h2020.symbiote.enabler.messaging.model.EnablerLogicDataAppearedMessage
 import eu.h2020.symbiote.enabler.messaging.model.PlatformProxyResourceInfo;
 import eu.h2020.symbiote.manager.AcquisitionManager;
 import eu.h2020.symbiote.security.TokenManager;
-import eu.h2020.symbiote.security.exceptions.aam.TokenValidationException;
-import eu.h2020.symbiote.security.token.Token;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.joda.time.DateTime;
-import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpMethod;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.client.RestTemplate;
 
 import java.util.ArrayList;
